fix: accommodate for new transform variable
